requirements.txt
anyio==4.4.0
arabic-reshaper==3.0.0
certifi==2024.7.4
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
emoji==2.12.1
fonttools==4.53.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
nltk==3.8.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
regex==2024.5.15
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
wordcloud==1.9.3
